% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{buttbs}
\alias{buttbs}
\title{Butterworth bandstop (notch) filter}
\usage{
buttbs(x, fs, low = 59, high = 61, order = 4L)
}
\arguments{
\item{x}{Numeric vector.}

\item{fs}{Sampling frequency (Hz).}

\item{low}{Lower cutoff (Hz, default 59).}

\item{high}{Upper cutoff (Hz, default 61).}

\item{order}{Filter order (default 4).}
}
\value{
List with filtered signal and metadata.
}
\description{
Butterworth bandstop (notch) filter
}
