% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{buttlp}
\alias{buttlp}
\title{Butterworth lowpass filter}
\usage{
buttlp(x, fs, cutoff, order = 4L)
}
\arguments{
\item{x}{Numeric vector.}

\item{fs}{Sampling frequency (Hz).}

\item{cutoff}{Cutoff frequency (Hz).}

\item{order}{Filter order (default 4).}
}
\value{
List with filtered signal and metadata.
}
\description{
Butterworth lowpass filter
}
