% Generated by moirais generate_rd.py
\name{canonicalize_cpads_data}
\alias{canonicalize_cpads_data}
\title{Canonicalize raw CPADS PUMF columns}
\description{
  Canonicalize raw CPADS PUMF columns
}
\usage{
  canonicalize_cpads_data(data)
}
\arguments{
  \item{data}{Raw CPADS data frame.}
}
\value{
  Data frame with canonical MOIRAIS analysis columns.
}
