% Generated by moirais generate_rd.py
\name{cohens_d}
\alias{cohens_d}
\title{Cohen's d effect size}
\description{
  Cohen's d effect size
}
\usage{
  cohens_d(x1, x2, pooled)
}
\arguments{
  \item{x1}{Numeric vector (group 1).}
  \item{x2}{Numeric vector (group 2).}
  \item{pooled}{Use pooled SD (default 'TRUE'). If 'FALSE', uses 'sd(x2)'.}
}
\value{
  Numeric Cohen's d.
}
