% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{hfd}
\alias{hfd}
\title{Higuchi fractal dimension}
\usage{
hfd(x, kmax = 10L)
}
\arguments{
\item{x}{Numeric vector.}

\item{kmax}{Maximum k (default 10).}
}
\value{
List with fractal dimension value.
}
\description{
Higuchi fractal dimension
}
