% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{hurst_r}
\alias{hurst_r}
\title{Hurst exponent via R/S analysis}
\usage{
hurst_r(x)
}
\arguments{
\item{x}{Numeric vector (time series).}
}
\value{
List with H exponent and interpretation.
}
\description{
Hurst exponent via R/S analysis
}
