% Generated by moirais generate_rd.py
\name{moirais_paths}
\alias{moirais_paths}
\title{Find a project root directory}
\description{
  Resolve standard project paths
}
\usage{
  moirais_paths(project_root)
}
\arguments{
  \item{project_root}{Project root directory. If 'NULL', inferred from the}
}
\value{
  Named list of key paths.
}
