% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{pcg_filter}
\alias{pcg_filter}
\title{PCG bandpass filter}
\usage{
pcg_filter(x, fs = 2000, low = 25, high = 400)
}
\arguments{
\item{x}{Numeric vector (PCG signal).}

\item{fs}{Sampling frequency (Hz, default 2000).}

\item{low}{Lower cutoff (Hz, default 25).}

\item{high}{Upper cutoff (Hz, default 400).}
}
\value{
List with filtered signal.
}
\description{
PCG bandpass filter
}
