% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{sgolay_smooth}
\alias{sgolay_smooth}
\title{Savitzky-Golay smoothing filter}
\usage{
sgolay_smooth(x, window_length = 11L, polyorder = 3L)
}
\arguments{
\item{x}{Numeric vector.}

\item{window_length}{Window length (odd integer, default 11).}

\item{polyorder}{Polynomial order (default 3).}
}
\value{
List with filtered signal.
}
\description{
Savitzky-Golay smoothing filter
}
